Extract UI initialization from main
